test: reorganize and consolidate test files by module
